Update(e2e): Move fixtures to common folder for sharing between tests.
