--- libs/e2e/src/lib/utils/users.ts
@@ -1,4 +1,0 @@
-import { User } from './type';
-import userFixture from 'tools/fixtures/users.json';
-
-export const USERS: Partial<User>[] = userFixture;
